Add petIdSchema and update editPet and deletePet functions

+++ src/actions/actions.ts
@@ -6,13 +6,14 @@ import { PetEssentials } from '@/lib/types';
 
 import { sleep } from '@/lib/utils';
 import { revalidatePath } from 'next/cache';
-import { petFormSchema } from '@/lib/validations';
+import { petFormSchema, petIdSchema } from '@/lib/validations';
 
 // When we get data from the client, we are not sure what it will be, so we use the unknown type.
 export async function addPet(pet: unknown) {
   await sleep(1000);
 
   const validatedPet = petFormSchema.safeParse(pet);
+
   if (!validatedPet.success) {
     return {
       message: 'Invalid pet data',
@@ -34,13 +35,22 @@ export async function addPet(pet: unknown) {
   revalidatePath('/app', 'layout');
 }
 
-export async function editPet(petId: Pet['id'], newPetData: PetEssentials) {
+export async function editPet(petId: unknown, newPetData: unknown) {
   await sleep(1000);
+
+  const validatedPetId = petIdSchema.safeParse(petId);
+  const validatedPet = petFormSchema.safeParse(newPetData);
+
+  if (!validatedPet.success || !validatedPetId.success) {
+    return {
+      message: 'Invalid pet data',
+    };
+  }
 
   try {
     await prisma.pet.update({
-      where: { id: petId },
-      data: newPetData,
+      where: { id: validatedPetId.data },
+      data: validatedPet.data,
     });
   } catch (error) {
     return {
@@ -51,12 +61,20 @@ export async function editPet(petId: Pet['id'], newPetData: PetEssentials) {
   revalidatePath('/app', 'layout');
 }
 
-export async function deletePet(petId: Pet['id']) {
+export async function deletePet(petId: unknown) {
   await sleep(1000);
+
+  const validatedPetId = petIdSchema.safeParse(petId);
+
+  if (!validatedPetId.success) {
+    return {
+      message: 'Invalid pet data',
+    };
+  }
 
   try {
     await prisma.pet.delete({
-      where: { id: petId },
+      where: { id: validatedPetId.data },
     });
   } catch (error) {
     return {
